Support DirEntry paths when checking is_fake_path
When using shutil.copytree from a non-overlayed dir to an overlayed dir,
copytree's implementation uses os.scandir that "returns an iterator of
os.DirEntry objects corresponding to the entries in the directory given
by path".

The os.DirEntry does not support calling startswith on it directly, so
this work handles that case by using it's name attribute instead.